Add doc.go
